Move InceptionBlock4 to components/cnn_convolutions.py
